Rename: move directory, temp annotation
