--- src/main/java/com/example/demo/config/secret/Secret.java
@@ -1,8 +1,0 @@
-package com.example.demo.config.secret;
-
-// TODO: 해당 KEY 값들을 꼭 바꿔서 사용해주세요!
-// TODO: .gitignore에 추가하는거 앚지 마세요!
-public class Secret {
-    public static String JWT_SECRET_KEY = "UwKYibQQgkW7g-*k.ap9kje-wxBHb9wdXoBT4vnt4P3sJWt-Nu";
-  
-}
