+++ src/main/java/com/example/demo/global/exception/BaseResponseStatus.java
@@ -1,4 +1,4 @@
-package com.example.demo.config;
+package com.example.demo.global.exception;
 
 import lombok.Getter;
