+++ src/main/java/com/example/demo/src/user/UserController.java
@@ -1,18 +1,15 @@
 package com.example.demo.src.user;
 
-import com.example.demo.config.BaseException;
-import com.example.demo.config.BaseResponse;
+import com.example.demo.global.exception.BaseException;
+import com.example.demo.global.exception.BaseResponse;
 import com.example.demo.src.user.model.*;
 import com.example.demo.utils.JwtService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import static com.example.demo.config.BaseResponseStatus.POST_USERS_EMPTY_EMAIL;
-import static com.example.demo.config.BaseResponseStatus.POST_USERS_INVALID_EMAIL;
+import static com.example.demo.global.exception.BaseResponseStatus.POST_USERS_EMPTY_EMAIL;
+import static com.example.demo.global.exception.BaseResponseStatus.POST_USERS_INVALID_EMAIL;
 import static com.example.demo.utils.ValidationRegex.isRegexEmail;
