+++ src/main/java/com/example/demo/src/user/UserService.java
@@ -1,7 +1,7 @@
 package com.example.demo.src.user;
 
 
-import com.example.demo.config.BaseException;
+import com.example.demo.global.exception.BaseException;
 
 import com.example.demo.src.user.model.PostUserReq;
 import com.example.demo.src.user.model.PostUserRes;
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import static com.example.demo.config.BaseResponseStatus.*;
+import static com.example.demo.global.exception.BaseResponseStatus.*;
 
 @RequiredArgsConstructor
 @Service
